[ical4j 4.x] Unify helpers / deduplicate (#262)
* Remove `Instant.toLocalDate()` from `DateUtils`

Use the more generic `Temporal.toLocalDate()` from `TimeApiExtensions` instead.

* Remove `Temporal.toEpochMilli` from `DateUtils`

Use `Temporal.toTimestamp()` from `AndroidTemporalMapper` instead. This function retains ical4j's old behavior when it comes to converting floating DATE and DATE-TIME values to a timestamp.

* Move functions from `AndroidTemporalMapper` to `AndroidTimeUtils`
