+++ lib/src/main/kotlin/at/bitfire/ical4android/util/DateUtils.kt
@@ -8,15 +8,6 @@ package at.bitfire.ical4android.util
 
 import net.fortuna.ical4j.model.TemporalAdapter
 import net.fortuna.ical4j.model.property.DateProperty
-import java.time.Instant
-import java.time.LocalDate
-import java.time.LocalDateTime
-import java.time.OffsetDateTime
-import java.time.ZoneId
-import java.time.ZoneOffset
-import java.time.ZonedDateTime
-import java.time.temporal.ChronoField
-import java.time.temporal.ChronoUnit
 import java.time.temporal.Temporal
 
 /**
@@ -56,40 +47,5 @@ object DateUtils {
      */
     fun isDateTime(date: Temporal?): Boolean =
         date != null && TemporalAdapter.isDateTimePrecision(date)
-
-    /**
-     * Converts the given [Instant] by truncating it to days, and converting into [LocalDate] by its
-     * epoch timestamp.
-     */
-    fun Instant.toLocalDate(): LocalDate {
-        val epochSeconds = truncatedTo(ChronoUnit.DAYS).epochSecond
-        return LocalDate.ofEpochDay(epochSeconds / (24 * 60 * 60 /*seconds in a day*/))
-    }
-
-    /**
-     * Converts the given generic [Temporal] into milliseconds since epoch.
-     * @param fallbackTimezone Any specific timezone to use as fallback if there's not enough
-     * information on the [Temporal] type (local types). Defaults to UTC.
-     * @throws IllegalArgumentException if the [Temporal] is from an unknown time, which also doesn't
-     * support [ChronoField.INSTANT_SECONDS]
-     */
-    fun Temporal.toEpochMilli(fallbackTimezone: ZoneId? = null): Long {
-        // If the temporal supports instant seconds, we can compute epoch millis directly from them
-        if (isSupported(ChronoField.INSTANT_SECONDS)) {
-            val seconds = getLong(ChronoField.INSTANT_SECONDS)
-            val nanos = get(ChronoField.NANO_OF_SECOND)
-            // Convert seconds and nanos to millis
-            return (seconds * 1000) + (nanos / 1_000_000)
-        }
-
-        return when (this) {
-            is Instant -> this.toEpochMilli()
-            is ZonedDateTime -> this.toInstant().toEpochMilli()
-            is OffsetDateTime -> this.toInstant().toEpochMilli()
-            is LocalDate -> this.atStartOfDay(fallbackTimezone ?: ZoneOffset.UTC).toInstant().toEpochMilli()
-            is LocalDateTime -> this.atZone(fallbackTimezone ?: ZoneOffset.UTC).toInstant().toEpochMilli()
-            else -> throw IllegalArgumentException("${this::class.java.simpleName} cannot be converted to epoch millis.")
-        }
-    }
 
 }
